RMSRCBUILD-268 Added JiraRestClientProvider
