+++ ReleaseProcessAutomation/Jira/Jira.cs
@@ -2,6 +2,7 @@ using System;
 using ReleaseProcessAutomation.Configuration.Data;
 using ReleaseProcessAutomation.Jira.CredentialManagement;
 using ReleaseProcessAutomation.Jira.ServiceFacadeImplementations;
+using ReleaseProcessAutomation.Jira.Utility;
 using RestSharp.Authenticators;
 
 namespace ReleaseProcessAutomation.Jira;
@@ -11,44 +12,24 @@ public class Jira
 {
     private readonly Config _config;
     private readonly IJiraCredentialManager _jiraCredentialManager;
+    private readonly IJiraRestClientProvider _jiraRestClientProvider;
 
     private JiraVersionCreator? _versionCreator;
 
-    public IJiraVersionCreator VersionCreator => _versionCreator ??= new JiraVersionCreator(JiraRestClient);
+    public IJiraVersionCreator VersionCreator => _versionCreator ??= new JiraVersionCreator(_jiraRestClientProvider.GetJiraRestClient());
 
     private JiraVersionReleaser? _versionReleaser;
-    public IJiraVersionReleaser VersionReleaser => _versionReleaser ??= new JiraVersionReleaser(JiraRestClient);
+    public IJiraVersionReleaser VersionReleaser => _versionReleaser ??= new JiraVersionReleaser(_jiraRestClientProvider.GetJiraRestClient());
     
     public IJiraAuthenticationWrapper AuthenticationWrapper { get; }
-
-    private JiraRestClient? _jiraRestClient;
-    private JiraRestClient JiraRestClient
-    {
-        get
-        {
-            if (_jiraRestClient == null)
-            {
-                if (_config.Jira.UseNTLM)
-                {
-                    _jiraRestClient = new JiraRestClient(_config.Jira.JiraURL, new NtlmAuthenticator());
-                }
-                else
-                {
-                    var credentials = _jiraCredentialManager.GetCredential(_config.Jira.JiraURL);
-                    //var credentials = new Credentials { Username = "user", Password = "password" };
-                    _jiraRestClient = new JiraRestClient(_config.Jira.JiraURL, new HttpBasicAuthenticator(credentials.Username, credentials.Password));
-                }
-            }
-
-            return _jiraRestClient;
-        }
-    }
+    
 
 
-    public Jira (Config config, IJiraCredentialManager jiraCredentialManager, IJiraAuthenticationWrapper authenticationWrapper)
+    public Jira (Config config, IJiraCredentialManager jiraCredentialManager, IJiraAuthenticationWrapper authenticationWrapper, IJiraRestClientProvider jiraRestClientProvider)
     {
         _config = config;
         _jiraCredentialManager = jiraCredentialManager;
+        _jiraRestClientProvider = jiraRestClientProvider;
         AuthenticationWrapper = authenticationWrapper;
     }
 }
