+++ ReleaseProcessAutomation/Jira/Utility/IJiraRestClientProvider.cs
@@ -1,0 +1,8 @@
+using ReleaseProcessAutomation.Jira.ServiceFacadeImplementations;
+
+namespace ReleaseProcessAutomation.Jira.Utility;
+
+public interface IJiraRestClientProvider
+{
+  JiraRestClient GetJiraRestClient();
+}
